Fix ty type-check errors in new tests
The CI lint job runs `ty` type validation (which my local `scripts/fix`
couldn't reach). Seven errors across four packages:

- plain-scan, plain-tunnel: `from conftest import ...` isn't resolvable by
  ty (no repo test uses that pattern). Inline the small synthetic-response
  / client factories into each test file and drop the conftests.
- plain-auth: `self.user.username` in the WhoView test helper — the user
  type comes through plain.auth's TYPE_CHECKING stub, so annotate with a
  targeted `# ty: ignore[unresolved-attribute]` (login_required guarantees
  a user at runtime).
- plain-email: `# ty: ignore` on the intentional wrong-type `to=<str>`
  argument and on the dynamic `get_payload()` element attribute.

All four suites still pass and `ty check` is clean for each.
